# docs/openapi/agent.v1.json
{
  "components": {
    "schemas": {
...
      "AgentAttrGroupSchema": {
        "additionalProperties": false,
        "properties": {
          "attrs": {
            "items": {
              "$ref": "#/components/schemas/AgentAttrSchema"
            },
            "type": "array"
          },
          "groupName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "groupName",
          "attrs"
        ],
        "type": "object"
      },
      "AgentAttrSchema": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "unit": {
            "type": [
              "string",
              "null"
            ]
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "value",
          "unit"
        ],
        "type": "object"
      },
      "AgentErrorCodeSchema": {
        "enum": [
          0,
          40001,
          40010,
          40101,
          40102,
          40301,
          40401,
          40402,
          40403,
          40404,
          40501,
          40901,
          42901,
          50001
        ],
        "type": "number"
      },
      "AgentErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/AgentErrorCodeSchema"
          },
          "data": {},
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ],
        "type": "object"
      },
...
      "AgentProductSpecsSchema": {
        "additionalProperties": false,
        "properties": {
          "attrGroups": {
            "items": {
              "$ref": "#/components/schemas/AgentAttrGroupSchema"
            },
            "type": "array"
          },
          "brand": {
            "type": [
              "string",
              "null"
            ]
          },
          "categoryPath": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "type": "array"
          },
          "contentHash": {
            "minLength": 1,
            "type": "string"
          },
          "mainImage": {
            "type": [
              "string",
              "null"
            ]
          },
          "skus": {
            "items": {
              "$ref": "#/components/schemas/AgentSpecSkuSchema"
            },
            "type": "array"
          },
          "specDimensions": {
            "items": {
              "$ref": "#/components/schemas/AgentSpecDimensionSchema"
            },
            "type": "array"
          },
          "spuId": {
            "$ref": "#/components/schemas/UlidSchema"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatusSchema"
          },
          "subtitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/IsoDateTimeSchema"
          }
        },
        "required": [
          "spuId",
          "title",
          "subtitle",
          "brand",
          "categoryPath",
          "status",
          "mainImage",
          "updatedAt",
          "contentHash",
          "attrGroups",
          "specDimensions",
          "skus"
        ],
        "type": "object"
      },
      "AgentProductSpecsSuccessResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 0,
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/AgentProductSpecsSchema"
          },
          "message": {
            "const": "ok",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ],
        "type": "object"
      },
      "AgentProductStockSchema": {
        "additionalProperties": false,
        "properties": {
          "available": {
            "type": "boolean"
          },
          "checkQuantity": {
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "type": "integer"
          },
          "shipFrom": {
            "items": {
              "$ref": "#/components/schemas/StockShipFromSchema"
            },
            "type": "array"
          },
          "skus": {
            "items": {
              "$ref": "#/components/schemas/AgentStockSkuSchema"
            },
            "type": "array"
          },
          "spuId": {
            "$ref": "#/components/schemas/UlidSchema"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatusSchema"
          },
          "totalStock": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/IsoDateTimeSchema"
          }
        },
        "required": [
          "spuId",
          "status",
          "checkQuantity",
          "available",
          "totalStock",
          "updatedAt",
          "shipFrom",
          "skus"
        ],
        "type": "object"
      },
      "AgentProductStockSuccessResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 0,
            "type": "number"
          },
          "data": {
            "$ref": "#/components/schemas/AgentProductStockSchema"
          },
          "message": {
            "const": "ok",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ],
        "type": "object"
      },
...
      "AgentSpecDimensionSchema": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/AnonSchema9ac136ed"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "name",
          "values"
        ],
        "type": "object"
      },
      "AgentSpecSkuSchema": {
        "additionalProperties": false,
        "properties": {
          "inStock": {
            "type": "boolean"
          },
          "marketPrice": {
            "$ref": "#/components/schemas/MoneySchema"
          },
          "price": {
            "$ref": "#/components/schemas/MoneySchema"
          },
          "skuCode": {
            "minLength": 1,
            "type": "string"
          },
          "skuId": {
            "$ref": "#/components/schemas/UlidSchema"
          },
          "spec": {
            "$ref": "#/components/schemas/SkuSpecSchema"
          },
          "status": {
            "$ref": "#/components/schemas/SkuStatusSchema"
          }
        },
        "required": [
          "skuId",
          "skuCode",
          "spec",
          "price",
          "marketPrice",
          "status",
          "inStock"
        ],
        "type": "object"
      },
      "AgentStockSkuSchema": {
        "additionalProperties": false,
        "properties": {
          "inStock": {
            "type": "boolean"
          },
          "restockEta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IsoDateSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "skuCode": {
            "minLength": 1,
            "type": "string"
          },
          "skuId": {
            "$ref": "#/components/schemas/UlidSchema"
          },
          "spec": {
            "$ref": "#/components/schemas/SkuSpecSchema"
          },
          "stock": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "skuId",
          "skuCode",
          "spec",
          "stock",
          "inStock",
          "restockEta"
        ],
        "type": "object"
      },
...
      "AnonSchema9ac136ed": {
        "minLength": 1,
        "type": "string"
      },
...
      "IsoDateSchema": {
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
        "type": "string"
      },
      "IsoDateTimeSchema": {
        "format": "date-time",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
        "type": "string"
      },
...
      "MoneySchema": {
        "maximum": 9007199254740991,
        "minimum": 0,
        "type": "integer"
      },
...
      "ProductStatusSchema": {
        "enum": [
          "draft",
          "pending_review",
          "onsale",
          "offsale",
          "rejected"
        ],
        "type": "string"
      },
...
      "SkuSpecSchema": {
        "additionalProperties": {
          "type": "string"
        },
        "propertyNames": {
          "type": "string"
        },
        "type": "object"
      },
      "SkuStatusSchema": {
        "enum": [
          "active",
          "inactive"
        ],
        "type": "string"
      },
      "StockShipFromSchema": {
        "additionalProperties": false,
        "properties": {
          "city": {
            "minLength": 1,
            "type": "string"
          },
          "province": {
            "minLength": 1,
            "type": "string"
          },
          "storeId": {
            "$ref": "#/components/schemas/UlidSchema"
          },
          "storeName": {
            "minLength": 1,
            "type": "string"
          },
          "supportsPickup": {
            "type": "boolean"
          },
          "type": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "storeId",
          "storeName",
          "type",
          "city",
          "province",
          "supportsPickup"
        ],
        "type": "object"
      },
      "UlidSchema": {
        "maxLength": 26,
        "minLength": 26,
        "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$",
        "type": "string"
      }
    },
    "securitySchemes": {
      "ServiceToken": {
        "description": "服务令牌（07 §7.8.1）：**非** Bearer，**不允许**放 query string。明文仅创建时返回一次；服务端存 `HMAC-SHA256(AGENT_TOKEN_PEPPER, token)`。",
        "in": "header",
        "name": "X-Service-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "PiEcho Agent 只读接口（GET-only，六端点）。契约版本载体：请求头 `X-Contract-Version`（缺失视为 `1`，不报错）。\n权威来源：`packages/shared/src/contracts/agent.ts`（Zod Schema 为唯一真相）。\n本文件由 `scripts/export-openapi.ts` 生成，**请勿手改**。",
    "title": "DShop Agent API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/agent/products/{spuId}/specs": {
      "get": {
        "operationId": "getAgentProductSpecs",
        "parameters": [
          {
            "description": "商品 SPU 的 26 位 ULID",
            "in": "path",
            "name": "spuId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UlidSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProductSpecsSuccessResponse"
                }
              }
            },
            "description": "成功（07 §7.1：HTTP 200、`code = 0`、`message = \"ok\"`；`data` 为下表端点 Schema）"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40001 参数校验失败；40010 不支持的契约版本"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40101 服务令牌缺失或无效；40102 服务令牌已吊销或已过期"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40301 令牌 scope 不含该资源"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40401 订单不存在；40402 商品不存在或已删除；40403 售后单不存在；40404 政策分类无生效条款"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40501 对 Agent 组使用了非 GET 方法"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40901 受控写幂等冲突或状态不允许"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 42901 触发限流"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 50001 服务内部错误"
          }
        },
        "security": [
          {
            "ServiceToken": []
          }
        ],
        "summary": "查询商品规格与参数白皮书（**不含**库存数值，仅 `inStock`）",
        "tags": [
          "agent"
        ],
        "x-agent-scope": "agent:product:read",
        "x-burst": 20,
        "x-cache-ttl-seconds": 60,
        "x-rate-limit-per-min": 300
      }
    },
    "/api/v1/agent/products/{spuId}/stock": {
      "get": {
        "operationId": "getAgentProductStock",
        "parameters": [
          {
            "description": "商品 SPU 的 26 位 ULID",
            "in": "path",
            "name": "spuId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UlidSchema"
            }
          },
          {
            "description": "限定单个 SKU（26 位 ULID）",
            "in": "query",
            "name": "skuId",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UlidSchema"
            }
          },
          {
            "description": "目标购买数量，默认 1（用于计算 `available`）",
            "in": "query",
            "name": "quantity",
            "required": false,
            "schema": {
              "default": 1,
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "description": "收货地区码（**预留**：多仓就近判断，一期不使用）",
            "in": "query",
            "name": "regionCode",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProductStockSuccessResponse"
                }
              }
            },
            "description": "成功（07 §7.1：HTTP 200、`code = 0`、`message = \"ok\"`；`data` 为下表端点 Schema）"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40001 参数校验失败；40010 不支持的契约版本"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40101 服务令牌缺失或无效；40102 服务令牌已吊销或已过期"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40301 令牌 scope 不含该资源"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40401 订单不存在；40402 商品不存在或已删除；40403 售后单不存在；40404 政策分类无生效条款"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40501 对 Agent 组使用了非 GET 方法"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 40901 受控写幂等冲突或状态不允许"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 42901 触发限流"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentErrorResponse"
                }
              }
            },
            "description": "失败：统一错误信封 `{code,message,data}` —— 50001 服务内部错误"
          }
        },
        "security": [
          {
            "ServiceToken": []
          }
        ],
        "summary": "查询商品可售库存与 SKU 级预计到货时间",
        "tags": [
          "agent"
        ],
        "x-agent-scope": "agent:product:read",
        "x-burst": 20,
        "x-cache-ttl-seconds": 30,
        "x-rate-limit-per-min": 300
      }
    }
  },
  "servers": [
    {
      "description": "演示/生产环境（docs/07 §7.2 Base URL 的 origin）",
      "url": "https://api.dshop.example.com"
    },
    {
      "description": "本地 `wrangler dev`（apps/api/wrangler.jsonc）",
      "url": "http://localhost:8787"
    }
  ],
...
}